Refactoring the progress_section.dart
